renamed text_editor to model and reflected rename across modules.
